+++ pkg/googleAPI/client.go
@@ -1,0 +1,79 @@
+package googleAPI
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"golang.org/x/oauth2"
+	"golang.org/x/oauth2/google"
+	"google.golang.org/api/calendar/v3"
+	"io/ioutil"
+	"net/http"
+	"os"
+)
+
+
+// GetClient
+func GetClient() (*http.Client, error) {
+	// credentials.jsonを読み込み
+	b, err := ioutil.ReadFile("../../config/credentials.json")
+	if err != nil {
+		return nil, fmt.Errorf("Unable to read client secret file: %v", err)
+	}
+
+	// クライアント作成
+	config, err := google.ConfigFromJSON(b, calendar.CalendarReadonlyScope)
+	if err != nil {
+		return nil, fmt.Errorf("Unable to parse client secret file to cofing: %v", err)
+	}
+
+	tokenFile := "../../config/token.json"
+	token, err := tokenFromFile(tokenFile)
+	if err != nil {
+		token, err = getTokenFromWeb(config)
+		if err != nil {
+			return nil, err
+		}
+		saveToken(tokenFile, token)
+	}
+	return config.Client(context.Background(), token), nil
+}
+
+func tokenFromFile(fileName string) (*oauth2.Token, error) {
+	file, err := os.Open(fileName)
+	if err != nil {
+		return nil, err
+	}
+	defer file.Close()
+	token := &oauth2.Token{}
+	err = json.NewDecoder(file).Decode(token)
+	return token, err
+}
+
+func getTokenFromWeb(config *oauth2.Config) (*oauth2.Token, error) {
+	authURL := config.AuthCodeURL("state-token", oauth2.AccessTypeOffline)
+
+	fmt.Printf("Go to the following link in your browser then type the "+
+		"authorization code: \n%v\n", authURL)
+	var authCode string
+
+	if _, err := fmt.Scan(&authCode); err != nil {
+		return nil, fmt.Errorf("Unable to read authorization code: %v", err)
+	}
+
+	token, err := config.Exchange(context.TODO(), authCode)
+	if err != nil {
+		return nil, fmt.Errorf("Unable to retrieve token from web: %v", err)
+	}
+	return token, nil
+}
+
+func saveToken(path string, token *oauth2.Token) error {
+	file, err := os.OpenFile(path, os.O_RDWR|os.O_CREATE|os.O_TRUNC, 0600)
+	if err != nil {
+		return fmt.Errorf("Unable to cache oauth token: %v", err)
+	}
+	defer file.Close()
+	json.NewEncoder(file).Encode(token)
+	return nil
+}
